% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cd_find_ids.R
\name{cd_find_ids}
\alias{cd_find_ids}
\title{Title: Find Specific Codes in Set of Columns}
\usage{
cd_find_ids(dat, cols, codes)
}
\arguments{
\item{dat}{data object to search}

\item{cols}{columns to search}

\item{codes}{codes to search for}
}
\value{
a tibble with a column for the codes found and a column for whether any codes were found
}
\description{
Title: Find Specific Codes in Set of Columns
}
\examples{
dat <- tibble(
  id = 1:3,
  code1 = c("A", "B", "C"),
  code2 = c("B", "C", "D")
)
cd_find_ids(dat, c("code1", "code2"), c("A", "B"))
# A tibble: 3 x 2
# codes_found any_codes
# <list>      <lgl>
# 1 <chr [1]>   TRUE
# 2 <chr [1]>   TRUE
# 3 <chr [0]>   FALSE
}
